requirements.txt
streamlit>=1.28.0
openai>=1.3.0
google-generativeai>=0.3.0
google-genai>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
